backend/requirements.txt
boto3
openpyxl
requests
urllib
json
python-dotenv